requirements.txt
python-telegram-bot
psycopg2
requests
